[#6587] move abstract ejb
